Usuario y TipoUsuario CRUD

+++ demodbd/src/main/java/com/example/dbd/demodbd/repositories/TipoUsuarioRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TipoUsuarioRepository extends CrudRepository<TipoUsuarioEntity, Long> {
+public interface TipoUsuarioRepository extends CrudRepository<TipoUsuarioEntity, Integer> {
 }
